+++ packages/plugin-sales-api/src/automations/index.ts
@@ -66,7 +66,7 @@ export default {
     subdomain,
     data: { target, config, relatedValueProps }
   }) => {
-    const models = generateModels(subdomain);
+    const models = await generateModels(subdomain);
 
     return await replacePlaceHolders({
       models,
@@ -75,12 +75,15 @@ export default {
       actionData: config,
       target: {
         ...target,
-        ['createdBy.department']: null,
-        ['createdBy.branch']: null,
-        ['createdBy.phone']: null,
-        ['customers.email']: null,
-        ['customers.phone']: null,
-        ['customers.fullName']: null
+        ['createdBy.department']: '-',
+        ['createdBy.branch']: '-',
+        ['createdBy.phone']: '-',
+        ['createdBy.email']: '-',
+        ['customers.email']: '-',
+        ['customers.phone']: '-',
+        ['customers.fullName']: '-',
+        link: '-',
+        pipelineLabels: '-'
       },
       relatedValueProps,
       complexFields: ['productsData']