+++ packages/plugin-tickets-api/src/fieldUtils.ts
@@ -300,6 +300,18 @@ export const generateFields = async ({ subdomain, data }) => {
       ...fields,
       {
         _id: Math.random(),
+        name: 'createdBy.email',
+        label: 'Created by Email',
+        type: 'String'
+      },
+      {
+        _id: Math.random(),
+        name: 'createdBy.fullName',
+        label: 'Created by Full Name',
+        type: 'String'
+      },
+      {
+        _id: Math.random(),
         name: 'createdBy.phone',
         label: 'Created by Phone',
         type: 'String'
@@ -332,6 +344,24 @@ export const generateFields = async ({ subdomain, data }) => {
         _id: Math.random(),
         name: 'customers.fullName',
         label: 'Customers FullName',
+        type: 'String'
+      },
+      {
+        _id: Math.random(),
+        name: 'link',
+        label: 'Link',
+        type: 'String'
+      },
+      {
+        _id: Math.random(),
+        name: 'pipelineLabels',
+        label: 'Pipeline Labels',
+        type: 'String'
+      },
+      {
+        _id: Math.random(),
+        name: 'branches.title',
+        label: 'Branches title',
         type: 'String'
       }
     ];